update: built a module from the FtApi class
